Add Comment repository, service & controller functionality. Add basic validation for DTO properties. Added query parameters for Motorcycle GetAll ednpoint. General refactoring

+++ api/Controllers/MotorcyclesController.cs
@@ -1,4 +1,5 @@
 ﻿using api.DTOs.Motorcycle;
+using api.Helpers.Queries;
 using api.Services.Contracts;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,16 +17,20 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] MotorcycleQuery query)
         {
-            var getDtos = await _motorcyclesService.GetAllAsync();
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
+            var getDtos = await _motorcyclesService.GetAllAsync(query);
 
             return Ok(getDtos);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var getDto = await _motorcyclesService.GetByIdAsync(id);
 
             return getDto == null ? NotFound() : Ok(getDto);
@@ -34,23 +39,29 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] MotorcyclePostDTO postDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var getDto = await _motorcyclesService.CreateAsync(postDto);
 
             return getDto == null ? NotFound() : Ok(getDto);
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id,
                                                 [FromBody] MotorcyclePutDTO putDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var getDto = await _motorcyclesService.UpdateAsync(id, putDto);
 
             return getDto == null ? NotFound() : Ok(getDto);
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var getDto = await _motorcyclesService.DeleteAsync(id);
 
             return getDto == null ? NotFound() : NoContent();
